Fix unsigned/signed comparison.

+++ src/core/Core.cpp
@@ -6632,7 +6632,7 @@ static float _AvgFrameDiff = 150.f;
 		{
 			if (EERIEMouseButton & 1)
 			{
-				if ((ARX_FLARES_Block==0) && (CurrSlot<MAX_SLOT)) 
+				if ((ARX_FLARES_Block==0) && (CurrSlot<(long)MAX_SLOT)) 
 					ARX_SPELLS_AddPoint(&DANAEMouse);
 				else
 				{
